Move to dep on a newer hadoop for mr2 API's for FS. Fixes up the hadoop platform test
