+++ scalding-core/src/main/scala/com/twitter/scalding/filecache/DistributedCacheFile.scala
@@ -6,7 +6,7 @@ import java.io.File
 import java.net.URI
 import java.nio.ByteBuffer
 import org.apache.hadoop.conf.Configuration
-import org.apache.hadoop.filecache.{ DistributedCache => HDistributedCache }
+import org.apache.hadoop.mapreduce.filecache.{ DistributedCache => HDistributedCache }
 import org.apache.hadoop.fs.Path
 
 object URIHasher {
